app/backend/requirements.txt
python-dotenv
pydantic
requests
google-api-python-client
tiktoken
mistralai>=0.0.11

fastapi
uvicorn
python-multipart
pypdf
markdown
PyMuPDF
boto3

docling==2.15.1
docling-core==2.15.1
docling-ibm-models==3.2.1
docling-parse==3.1.1
pypdfium2==4.30.0

torch==2.6.0
torchvision==0.21.0

litellm==1.63.3
google-generativeai
pinecone[grpc]
langchain
chromadb
nltk
sentence_transformers

selenium
webdriver-manager
beautifulsoup4